+++ wireguard-go-rust-wrapper/wireguard-go/main.go
@@ -31,10 +31,10 @@ import (
 	"time"
 	"unsafe"
 
-	"golang.zx2c4.com/wireguard/conn"
-	"golang.zx2c4.com/wireguard/device"
-	"golang.zx2c4.com/wireguard/ipc"
-	"golang.zx2c4.com/wireguard/tun"
+	"github.com/NordSecurity/wireguard-go/conn"
+	"github.com/NordSecurity/wireguard-go/device"
+	"github.com/NordSecurity/wireguard-go/ipc"
+	"github.com/NordSecurity/wireguard-go/tun"
 )
 
 type CLogger struct {