+++ wireguard-go-rust-wrapper/wireguard-go/start_with_tun_std.go
@@ -1,4 +1,4 @@
-// +build !linux
+//go:build !linux
 
 package main
 
@@ -17,5 +17,5 @@ package main
 import "C"
 
 func PlatformSpecific_StartWithTun(fd C.int32_t, log C.wg_go_log_cb) C.int32_t {
-  return -1
+	return -1
 }