+++ changelog.md
@@ -1,3 +1,11 @@
+### v4.2.9
+---
+* LLT-5095: Send critical event to app when adapter is dead
+* LLT-5097: Update wireguard wrapper to use wireguard-go fork and fix uapi get to throw error when device dead
+* LLT-5096: Fix uapi get/set to throw error when device dead
+
+<br>
+
 ### v4.2.8
 ---
 * LLT-4644: Fix wintun adapter creation